+++ cancer_api/__init__.py
@@ -4,13 +4,14 @@ __init__.py
 Where it all begins.
 """
 
-from main import connect, get_connection, set_connection
-from connections import MysqlConnection, SqliteConnection
-from mutations import SingleNucleotideVariant, Indel, CopyNumberVariation, StructuralVariation
-from metadata import Patient, Sample, Library
-from annotations import Gene, Transcript, Exon, Protein
+from base import Session
+from connections import *
+from mutations import *
+from effects import *
+from metadata import *
+from annotations import *
 from files import *
 from parsers import *
 from utils import *
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"