+++ cancer_api/mutations.py
@@ -7,10 +7,11 @@ in cancer, notably SNVs, indels, CNVs and SVs.
 
 from sqlalchemy import Column, Integer, String, Text, Float, Enum, ForeignKey
 from sqlalchemy.orm import relationship
-from base import Base
+import base
+import misc
 
 
-class Mutation(Base):
+class Mutation(base.Base):
     """Base class for all mutations"""
 
     id = Column(Integer, primary_key=True)
@@ -21,6 +22,11 @@ class Mutation(Base):
     __mapper_args__ = {"polymorphic_on": mutation_type}
 
     library = relationship("Library", backref="mutations")
+
+    def is_overlap(self, chrom, pos1, pos2=None, margin=0):
+        """Return whether given position overlaps with mutation.
+        """
+        raise NotImplementedError()
 
 
 class SingleNucleotideVariant(Mutation):
@@ -34,7 +40,16 @@ class SingleNucleotideVariant(Mutation):
     ref_count = Column(Integer)
     alt_count = Column(Integer)
 
+    __mapper_args__ = {'polymorphic_identity': 'snv'}
+
     mutation = relationship("Mutation", backref="snv")
+
+    def is_overlap(self, chrom, pos1, pos2=None, margin=0):
+        """Return whether given position overlaps with SNV.
+        """
+        snv_interval = misc.GenomicInterval(self.chrom, self.pos)
+        query_interval = misc.GenomicInterval(chrom, pos1, pos2)
+        return snv_interval.is_overlap(query_interval, margin)
 
 
 class Indel(Mutation):
@@ -47,6 +62,8 @@ class Indel(Mutation):
     alt_allele = Column(Text)
     ref_count = Column(Integer)
     alt_count = Column(Integer)
+
+    __mapper_args__ = {'polymorphic_identity': 'indel'}
 
     mutation = relationship("Mutation", backref="indel")
 
@@ -62,8 +79,52 @@ class StructuralVariation(Mutation):
     pos2 = Column(Integer)
     strand2 = Column(String(length=1))
     sv_type = Column(Enum("translocation", "inversion", "insertion", "deletion", "duplication"))
+    t_ref_count = Column(Integer)
+    n_ref_count = Column(Integer)
+    t_alt_count = Column(Integer)
+    n_alt_count = Column(Integer)
+    t_ref_spanning_reads = Column(Integer)
+    n_ref_spanning_reads = Column(Integer)
+    t_ref_read_pairs = Column(Integer)
+    n_ref_read_pairs = Column(Integer)
+    t_alt_spanning_reads = Column(Integer)
+    n_alt_spanning_reads = Column(Integer)
+    t_alt_read_pairs = Column(Integer)
+    n_alt_read_pairs = Column(Integer)
+
+    __mapper_args__ = {'polymorphic_identity': 'sv'}
 
     mutation = relationship("Mutation", backref="sv")
+
+    def predict_effects(self, db_sess):
+        """Predict the effect of the SV
+        """
+        effects = []
+        # First, consider structural effects
+        if self.sv_type in ["translocation", "inversion"]:
+            pass
+        # Second, consider copy number effects
+        if self.sv_type in ["duplication", "deletion"]:
+            pass
+        # Return effects
+        return effects
+
+    def is_overlap(self, chrom, pos1, pos2=None, margin=0):
+        """Return whether the given position overlap with
+        the structural variation.
+        The margin defines how close the events can be to
+        be considered overlapping (e.g., within 10 bp).
+        """
+        query_interval = misc.GenomicInterval(chrom, pos1, pos2)
+        if self.chrom1 == self.chrom2:
+            sv_interval = misc.GenomicInterval(self.chrom1, self.pos1, self.pos2)
+            is_overlap = sv_interval.is_overlap(query_interval, margin)
+        else:
+            sv_interval1 = misc.GenomicInterval(self.chrom1, self.pos1)
+            sv_interval2 = misc.GenomicInterval(self.chrom2, self.pos2)
+            is_overlap = (sv_interval1.is_overlap(query_interval, margin) or
+                          sv_interval2.is_overlap(query_interval, margin))
+        return is_overlap
 
 
 class CopyNumberVariation(Mutation):
@@ -76,5 +137,7 @@ class CopyNumberVariation(Mutation):
     size = Column(Integer)
     fold_change = Column(Float)
     copy_state = Column(Integer)
+
+    __mapper_args__ = {'polymorphic_identity': 'cnv'}
 
     mutation = relationship("Mutation", backref="cnv")